requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
matplotlib
seaborn
shap
mlflow
python-dotenv
